bugfix: #170 fixing engine builder input params

+++ component_integration_tests/rust_test_scenarios/src/internals/helpers/runtime_helper.rs
@@ -6,7 +6,7 @@ use tracing::debug;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Runtime {
-    task_queue_size: usize,
+    task_queue_size: u32,
     workers: usize,
 }
